Merged PR 130200: bug - countries not accepting orders with repeated skus!
Thank you for your contribution to out B2B Data Mass Framework.

## Type of changes

- [x] Bug fix
- [ ] New feature
- [ ] Documentation / docstrings
- [ ] Tests
- [ ] Refactoring
- [ ] Other

## Checklist

- [x] I accept that @<[GHQ_B2B_Delta]\Data Mass> is required as a PR reviewer.
- [x] I guarantee that the proposed changes have been tested and are working without any side effects

### Documentation:

- [x] Jira ticket: https://ab-inbev.atlassian.net/browse/BEESDM-82
- [ ] Confluence page (If exists): https://ab-inbev.atlassian.net/wiki/spaces/PKB/overview

## Description

Please describe your changes here.

Countries are not accepting orders with repeated skus, so the solution
is create a set to guarantee unique skus and ask the user the quantity of
skus to be inserted in the order. Using click to validade the user entry!