Merged PR 159256: User option to input an owner to the account.
Thank you for your contribution to out B2B Data Mass Framework.

## Type of changes

- [ ] Bug fix
- [x] New feature
- [ ] Documentation / docstrings
- [ ] Tests
- [ ] Refactoring
- [ ] Other

## Checklist

- [x] I accept that @<[GHQ_B2B_Delta]\Data Mass> is required as a PR reviewer.
- [x] I guarantee that the proposed changes have been tested and are working without any side effects

### Documentation:

- [x] Jira ticket: https://ab-inbev.atlassian.net/browse/BEESDM-160
- [ ] Confluence page (If exists): https://ab-inbev.atlassian.net/wiki/spaces/PKB/overview

## Steps to reproduce
1 - 1 - 1 ->  create account
2 - 2 - 1 -> get values of created account (image2).

## Description

When the user creates an order via web_app it receives an e-mail with the infos
related to the order and it's status. The same behavior should happens via data_mass
but we sent an default value in the owner field, now we can input an e-mail to receive
something like image1, input the first and last name of the owner, as expected in MS.

![image1.png](https://ab-inbev.visualstudio.com/f9f2f3f2-8e28-4d73-9aaf-6554466d25f7/_apis/git/repositories/17efa418-3286-4df1-9941-80ca6447a8dd/pullRequests/159256/attachments/image.png =300x)
(image1)

![image2.png](https://ab-inbev.visualstudio.com/f9f2f3f2-8e28-4d73-9aaf-6554466d25f7/_apis/git/repositories/17efa418-3286-4df1-9941-80ca6447a8dd/pullRequests/159256/attachments/image%20%282%29.png =300x)
(image2)